requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi
charset-normalizer==2.0.6
cryptography==3.4.8
Django==3.2.7
django-filter==21.1
djangorestframework==3.12.4
idna==3.2
pyjwt
Pillow==8.3.2
pycparser==2.20
python-decouple==3.4
pytz==2021.1
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
psycopg2